requirements.txt
Flask
Flask-Cors
gunicorn
numpy
google-generativeai